requirements.txt
flask>=3,<4
python-dotenv>=1.0,<2
requests>=2.31,<3
Pillow>=10.0,<11
